+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to
 ### Added
 
 - Naive implementation of Yaz0 compression.
+- A `compressBuffer` function.
 
 ## [0.2.0] - 2020-09-13